pyproject.toml
[tool.poetry]
name = "computervision2024"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.2"
opencv-python = "^4.10.0.84"
matplotlib = "^3.9.2"
pillow = "^10.4.0"
ipykernel = "^6.29.5"
scipy = "^1.14.1"
scikit-learn = "^1.5.2"
plotly = "^5.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
